requirements.txt
streamlit
pandas
sentence-transformers
scikit-learn
torch
ftfy
regex
requests
Pillow
git+https://github.com/openai/CLIP.git
